move methods to LoopCharacteristicTrait
